Fix Users model structure
