--- backend/src/main/java/pk/pz/ultigrade/Views.java
@@ -1,7 +1,0 @@
-package pk.pz.ultigrade;
-
-public class Views {
-
-    public static class Public {}
-    public static class Internal extends Public {}
-}
