+++ backend/src/main/java/pk/pz/ultigrade/controllers/AuthController.java
@@ -5,6 +5,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import pk.pz.ultigrade.configurations.SecurityConfig;
+import pk.pz.ultigrade.models.InsertUserEntity;
+import pk.pz.ultigrade.models.RoleEntity;
+import pk.pz.ultigrade.models.UsersEntity;
+import pk.pz.ultigrade.repositories.InsertUserEntityRepository;
 import pk.pz.ultigrade.repositories.UserEntityRepository;
 import pk.pz.ultigrade.requests.RegisterRequest;
 import pk.pz.ultigrade.util.JsonResponse;
@@ -13,8 +17,9 @@ import pk.pz.ultigrade.util.JsonResponse;
 @CrossOrigin(value = "*", maxAge = 3600)
 @RequestMapping("/api/auth")
 public class AuthController {
-//    @Autowired
-//    UserEntityRepository userEntityRepository;
+
+    @Autowired
+    InsertUserEntityRepository insertUserEntityRepository;
 
     @Autowired
     PasswordEncoder encoder;
@@ -22,26 +27,21 @@ public class AuthController {
 
     @PostMapping(SecurityConfig.SIGNUP_PAGE)
     public ResponseEntity<?> signup(@RequestBody RegisterRequest registerRequest){
-//        if(userE/*ntityRepository.existsByPesel(registerRequest.getPesel())){
-//            return JsonResponse.badRequest("user with this pesel already exists!");
-//        }
-//
-//        if(!registerRequest.getPassword().equals(registerRequest.getConfirmedPassword()) ){
-//            return JsonResponse.badRequest("passwords dont match!");
-//        }*/
+        if(insertUserEntityRepository.existsByPesel(registerRequest.getPesel())){
+            return JsonResponse.badRequest("user with this pesel already exists!");
+        }
 
-//        UsersEntity usersEntity = new UsersEntity(
-//                registerRequest.getIdRole(),
-//            registerRequest.getName(),
-//            registerRequest.getSurname(),
-//            registerRequest.getPesel(),
-//            encoder.encode(registerRequest.getPassword()),
-//            registerRequest.getPesel(),
-//            registerRequest.getAdress(),
-//        "");
+        InsertUserEntity usersEntity = new InsertUserEntity(
+                registerRequest.getIdRole(),
+                registerRequest.getName(),
+                registerRequest.getSurname(),
+                encoder.encode(registerRequest.getPassword()),
+                registerRequest.getPesel(),
+                registerRequest.getAdress(),
+                ""
+        );
 
-
-//        userEntityRepository.save(usersEntity);
+        insertUserEntityRepository.save(usersEntity);
         return JsonResponse.ok("User Registrated!");
     }
