+++ backend/src/main/java/pk/pz/ultigrade/models/StudentGrades.java
@@ -1,7 +1,7 @@
 package pk.pz.ultigrade.models;
 
 import javax.persistence.Entity;
-
-@Entity
-public class StudentGrades {
-}
+//
+//@Entity
+//public class StudentGrades {
+//}
