+++ backend/src/main/java/pk/pz/ultigrade/models/UsersEntity.java
@@ -1,126 +1,14 @@
 package pk.pz.ultigrade.models;
 
-
-import javax.persistence.*;
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
 
 @Entity
-@Table(name = "users")
-public class UsersEntity {
+@DiscriminatorValue("not null")
+public class UsersEntity extends UsersBaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_user", unique = true, nullable = false)
-    private int idUser;
-    @Column(name = "id_role")
-    private int idRole;
-    @Column
-    private String name;
-    @Column
-    private String surname;
-    @Column
-    private String login;
-    @Column
-    private String password;
-    @Column
-    private String pesel;
-    @Column
-    private String adress;
-    @Column
-    private String phone;
-    @Column
-    private String email;
-
-    public UsersEntity(){}
-
-    public UsersEntity(int idRole, String name, String surname, String login, String password, String pesel, String adress, String phone, String email) {
-        this.idRole = idRole;
-        this.name = name;
-        this.surname = surname;
-        this.login = login;
-        this.password = password;
-        this.pesel = pesel;
-        this.adress = adress;
-        this.phone = phone;
-        this.email = email;
+    public UsersEntity(){
+        super();
     }
 
-    public int getIdUser() {
-        return idUser;
-    }
-
-    public void setIdUser(int idUser) {
-        this.idUser = idUser;
-    }
-
-    public int getIdRole() {
-        return idRole;
-    }
-
-    public void setIdRole(int idRole) {
-        this.idRole = idRole;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getPesel() {
-        return pesel;
-    }
-
-    public void setPesel(String pesel) {
-        this.pesel = pesel;
-    }
-
-    public String getAdress() {
-        return adress;
-    }
-
-    public void setAdress(String adress) {
-        this.adress = adress;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
 }
