+++ backend/src/main/java/pk/pz/ultigrade/repositories/InsertUserEntityRepository.java
@@ -1,0 +1,11 @@
+package pk.pz.ultigrade.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import pk.pz.ultigrade.models.InsertUserEntity;
+
+@Repository
+public interface InsertUserEntityRepository extends JpaRepository<InsertUserEntity, Integer> {
+
+    boolean existsByPesel(String pesel);
+}
