+++ backend/src/main/java/pk/pz/ultigrade/services/AuthUserDetailsService.java
@@ -5,6 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import pk.pz.ultigrade.details.UserDetailsImpl;
+import pk.pz.ultigrade.models.UsersBaseEntity;
 import pk.pz.ultigrade.models.UsersEntity;
 import pk.pz.ultigrade.repositories.UserEntityRepository;
 
@@ -17,7 +18,7 @@ public class AuthUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String login) throws UsernameNotFoundException {
-        Optional<UsersEntity> user = userEntityRepository.findByLogin(login);
+        Optional<UsersEntity> user = userEntityRepository.findByPesel(login);
         if(user.isPresent())
             return new UserDetailsImpl(user.get());
